requirements.txt
alembic
asyncpg
bcrypt
fastapi
python-dotenv
python-jose
sqlmodel
SQLAlchemy
uvicorn[standard]
certifi==2026.2.25
charset-normalizer==3.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.62.1
fsspec==2025.12.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
opencv-python==4.13.0.92
packaging==26.0
pillow==12.0.0
polars==1.39.2
polars-runtime-32==1.39.2
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.1
setuptools==70.2.0
six==1.17.0
sympy==1.14.0
torch==2.10.0+cpu
torchvision==0.25.0+cpu
typing_extensions==4.15.0
ultralytics==8.4.24
ultralytics-thop==2.0.18
urllib3==2.6.3
